+++ src/packages/interface.js
@@ -1,0 +1,11 @@
+class Interface {
+  constructor() {
+    this.components = new Set()
+  }
+
+  registerComponent({ component, container, position }) {
+    this.components.add({ component, container, position })
+  }
+}
+
+export default Interface
