+++ src/packages/test/Hello.vue
@@ -1,0 +1,9 @@
+<template>
+  <div>Hello</div>
+</template>
+
+<script>
+export default {
+  name: 'Hello',
+}
+</script>
